+++ src/EulerHS/Framework/Runtime.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE DerivingVia #-}
+
 module EulerHS.Framework.Runtime
   (
     -- * Framework Runtime
@@ -8,16 +10,26 @@ module EulerHS.Framework.Runtime
   , kvDisconnect
   , runPubSubWorker
   , shouldFlowLogRawSql
+  -- * Registration for self-signed certs
+  , CertificateRegistrationError(..)
+  , withSelfSignedFlowRuntime
   ) where
 
+import           Control.Monad.Trans.Except (throwE)
 import qualified Data.Map as Map (empty)
 import qualified Data.Pool as DP (destroyAllResources)
+import           Data.X509.CertificateStore (readCertificateStore)
 import qualified Database.Redis as RD
 import qualified EulerHS.Core.Runtime as R
 import qualified EulerHS.Core.Types as T
 import           EulerHS.Prelude
+import           Network.Connection (TLSSettings (TLSSettings))
 import           Network.HTTP.Client (Manager, newManager)
-import           Network.HTTP.Client.TLS (tlsManagerSettings)
+import           Network.HTTP.Client.TLS (mkManagerSettings, tlsManagerSettings)
+import           Network.TLS (ClientParams (clientShared, clientSupported),
+                              defaultParamsClient, sharedCAStore,
+                              supportedCiphers)
+import           Network.TLS.Extra.Cipher (ciphersuite_default)
 import           System.IO.Unsafe (unsafePerformIO)
 import qualified System.Mem as SYSM (performGC)
 
@@ -40,6 +52,60 @@ data FlowRuntime = FlowRuntime
   , _pubSubConnection         :: Maybe RD.Connection
   -- ^ Connection being used for Publish
   }
+
+-- | Possible issues that can arise when registering certificates.
+--
+-- @since 2.0.4.3
+newtype CertificateRegistrationError = NoCertificatesAtPath FilePath
+  deriving (Eq) via FilePath
+  deriving stock (Show)
+
+instance Exception CertificateRegistrationError
+
+-- | Works identically to 'withFlowRuntime', but takes an extra parameter. This
+-- parameter is a map of textual identifiers to paths where self-signed
+-- certificates can be found.
+--
+-- You can then use 'callAPI', providing 'Just' the textual identifier to use
+-- the self-signed certificate(s) provided.
+--
+-- The handler is provided an 'Either' to allow for graceful recovery; if you
+-- have nothing you can do with a 'CertificateRegistrationError', you can throw
+-- it.
+--
+-- @since 2.0.4.3
+withSelfSignedFlowRuntime ::
+  HashMap Text FilePath ->
+  Maybe (IO R.LoggerRuntime) ->
+  (Either CertificateRegistrationError FlowRuntime -> IO a) ->
+  IO a
+withSelfSignedFlowRuntime certPathMap mRTF handler = do
+  res <- runExceptT . traverse go $ certPathMap
+  case res of
+    Left err         -> handler . Left $ err
+    Right managerMap ->
+      bracket (fromMaybe R.createVoidLoggerRuntime mRTF) R.clearLoggerRuntime $
+        \loggerRT -> bracket (R.createCoreRuntime loggerRT) R.clearCoreRuntime $
+          \coreRT -> bracket (mkFlowRT coreRT managerMap) clearFlowRuntime $
+            handler . Right
+  where
+    go ::
+      FilePath ->
+      ExceptT CertificateRegistrationError IO Manager
+    go certPath = do
+      mCertStore <- lift . readCertificateStore $ certPath
+      case mCertStore of
+        Nothing    -> throwE . NoCertificatesAtPath $ certPath
+        Just store -> do
+          let defs = defaultParamsClient "localhost" ""
+          let clientParams = defs {
+            clientShared = (clientShared defs) { sharedCAStore = store },
+            clientSupported = (clientSupported defs) { supportedCiphers = ciphersuite_default }}
+          lift . newManager . mkManagerSettings (TLSSettings clientParams) $ Nothing
+    mkFlowRT :: R.CoreRuntime -> HashMap Text Manager -> IO FlowRuntime
+    mkFlowRT coreRT managers = do
+      frt <- createFlowRuntime coreRT
+      pure frt { _httpClientManagers = managers }
 
 -- | Create default FlowRuntime.
 createFlowRuntime :: R.CoreRuntime -> IO FlowRuntime