+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog for euler-hs
 
-## [2.0.4.3] - 
+## [2.1.0.0] - 2021-04-12
 
 * `ManagerSelector` is now a newtype.
+* Add `CertificateRegistrationError` data type.
+* Add 'withSelfSignedFlowRuntime` function for registering self-signed
+  certificates in a `FlowRuntime`.
 
 ## [2.0.4.2] - 2021-04-07